fix: FocusZone to pass ref to components wrapped in FZ
